Split the live-only PiP state observer into its own file (#78)
Patch coverage failed again, and my local check said it would pass. The
check was wrong: I measured coverage without `CI=true`, so playback
tests ran locally that CI skips. Under CI's actual conditions the diff
had 15 uncovered lines, not 6 — 77% against an 85% target.

The uncovered code is the whole observer body: the native-activity
transition and the timebase writes. Every branch is conditioned on the
player actively playing, and CI cannot drive libVLC to `.playing` at
all, so none of it is reachable there regardless of how it is tested.

Exempting `PiPController.swift` wholesale would have fixed the gate by
giving up enforcement on a large and mostly testable file. Moving the
observer into `PiPController+StateObserver.swift` and gating only that
keeps the rest enforced, and puts the genuinely unrunnable code
somewhere its name says so.

What stays measured is the part that can be: the retracking rules are
still a pure function with unit tests. What goes unmeasured is applying
those decisions to a live CMTimebase, not the decisions themselves.

`PiPController.swift` now contributes one added executable line to the
enforced patch, fully covered.
